+++ packages/jest-changed-files/src/hg.js
@@ -26,8 +26,8 @@ const adapter: SCMAdapter = {
       let args = ['status', '-amnu'];
       if (options && options.withAncestor) {
         args.push('--rev', 'ancestor(.^)');
-      } else if (options && options.toContributeTo) {
-        args.push('--rev', `ancestor(., ${options.toContributeTo})`);
+      } else if (options && options.changedSince) {
+        args.push('--rev', `ancestor(., ${options.changedSince})`);
       } else if (options && options.lastCommit === true) {
         args = ['tip', '--template', '{files%"{file}\n"}'];
       }