+++ packages/jest-cli/src/cli/args.js
@@ -24,7 +24,7 @@ export const check = (argv: Argv) => {
     'onlyChanged',
     'lastCommit',
     'changedFilesWithAncestor',
-    'changedFilesToContributeTo',
+    'changedSince',
   ]) {
     if (argv[key]) {
       argv.onlyChanged = true;
@@ -112,19 +112,19 @@ export const options = {
       ' dependency information.',
     type: 'string',
   },
-  changedFilesToContributeTo: {
+  changedFilesWithAncestor: {
+    description:
+      'Runs tests related to the current changes and the changes made in the ' +
+      'last commit. Behaves similarly to `--onlyChanged`.',
+    type: 'boolean',
+  },
+  changedSince: {
     description:
       'Runs tests related the changes since the provided branch. If the ' +
       'current branch has diverged from the given branch, then only changes ' +
       'made locally will be tested. Behaves similarly to `--onlyChanged`.',
     nargs: 1,
     type: 'string',
-  },
-  changedFilesWithAncestor: {
-    description:
-      'Runs tests related to the current changes and the changes made in the ' +
-      'last commit. Behaves similarly to `--onlyChanged`.',
-    type: 'boolean',
   },
   ci: {
     default: isCI,