+++ test_utils.js
@@ -13,8 +13,8 @@ import type {GlobalConfig, ProjectConfig} from 'types/Config';
 
 const DEFAULT_GLOBAL_CONFIG: GlobalConfig = {
   bail: false,
-  changedFilesToContributeTo: '',
   changedFilesWithAncestor: false,
+  changedSince: '',
   collectCoverage: false,
   collectCoverageFrom: [],
   collectCoverageOnlyFrom: null,