+++ types/Config.js
@@ -77,7 +77,7 @@ export type InitialOptions = {
   cacheDirectory?: Path,
   clearMocks?: boolean,
   changedFilesWithAncestor?: boolean,
-  changedFilesToContributeTo?: string,
+  changedSince?: string,
   collectCoverage?: boolean,
   collectCoverageFrom?: Array<Glob>,
   collectCoverageOnlyFrom?: {[key: string]: boolean},
@@ -160,7 +160,7 @@ export type SnapshotUpdateState = 'all' | 'new' | 'none';
 
 export type GlobalConfig = {|
   bail: boolean,
-  changedFilesToContributeTo: string,
+  changedSince: string,
   changedFilesWithAncestor: boolean,
   collectCoverage: boolean,
   collectCoverageFrom: Array<Glob>,